feat: bot init and refactor bot decorators
Now bot need to be initialised before registering message handlers.
Also created new functions to register wait_on actions in the database
