+++ examples/command_with_text_reply_handler/docker-compose.yml
@@ -1,0 +1,13 @@
+version: "3"
+services:
+  mongodb:
+    image: mongo
+
+  botstarter:
+    build: .
+    depends_on:
+      - mongodb
+    environment:
+      BOT_TOKEN: "${BOT_TOKEN}"
+      MONGODB_HOST: "mongodb"
+      DATABASE_NAME: "command_with_text_reply_handler"
